fix(servers): fail-closed audit + fold Slice 1 baseline into canonical test file
Review-fix round 1 (run c3749614):

- scripts/audit-server-tenant-scope.ts: require TURSO_DB_URL/
  TURSO_DB_AUTH_TOKEN explicitly and abort when absent, matching the existing
  Supabase check. The prior silent fallback to the app's dev-default
  file:./data/minion_hub.db let a missing/mistyped TURSO_DB_URL report a false
  PASS (empty local table -> zero mismatches) for a production readiness gate.
  Sharpened the TODO(handoff): the two-source audit still has not been run
  against non-production or production (no credentials in this sandbox), and
  that evidence plus the concrete re-key deployment record remain a human
  merge gate per the spec — an agent must not fabricate that output.

- Moved the updateServer Slice 1 baseline tests (same-tenant update,
  unknown-id no-op) and their in-memory filtering harness from the new
  server.service.update.test.ts into the existing
  src/server/services/server.service.test.ts, which is the file the spec's
  Slice 1/2 Definition of done names as the default focused gate. Deleted the
  split file so the focused-gate command actually exercises the new coverage.
