+++ src/server/services/server.service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { upsertServer, listServers, deleteServer } from './server.service';
+import { upsertServer, listServers, deleteServer, updateServer } from './server.service';
 import { createMockDb } from '$server/test-utils/mock-db';
 
 beforeEach(() => {
@@ -9,6 +9,37 @@ beforeEach(() => {
 vi.mock('$server/db/utils', () => ({
   newId: () => 'mock-server-id-00000001',
   nowMs: () => 1_700_000_000_000,
+}));
+
+// updateServer's Slice 1 baseline tests (specs/2026-08-18-hub-updateserver-tenant-scope-spec.md)
+// need a real filtering in-memory `servers` table instead of the createMockDb chain-recorder
+// above, which never inspects `.where()` predicates. Real column objects would need the real
+// @minion-stack/db/schema sqliteTable machinery, so the mock below only needs stable string
+// keys, mirrored by the drizzle-orm eq/and mock so `eq(servers.tenantId, x)` filters on the
+// `tenantId` property of a plain row object. createMockDb-based tests above are unaffected
+// since they ignore the actual column/predicate values passed through the chain.
+vi.mock('@minion-stack/db/schema', () => ({
+  servers: {
+    id: 'id',
+    tenantId: 'tenantId',
+    name: 'name',
+    url: 'url',
+    token: 'token',
+    tokenIv: 'tokenIv',
+    authMode: 'authMode',
+    lastConnectedAt: 'lastConnectedAt',
+    createdAt: 'createdAt',
+    updatedAt: 'updatedAt',
+  },
+  userServers: { userId: 'userId', serverId: 'serverId', createdAt: 'createdAt' },
+}));
+
+vi.mock('drizzle-orm', () => ({
+  eq: (col: string, val: unknown) => (r: Record<string, unknown>) => r[col] === val,
+  and:
+    (...fns: Array<(r: Record<string, unknown>) => boolean>) =>
+    (r: Record<string, unknown>) =>
+      fns.every((f) => f(r)),
 }));
 
 describe('upsertServer', () => {
@@ -69,5 +100,95 @@ describe('deleteServer', () => {
     const { db } = createMockDb();
     await deleteServer({ db, tenantId: 't1' }, 's1');
     expect(db.delete).toHaveBeenCalled();
+  });
+});
+
+// Slice 1 baseline (specs/2026-08-18-hub-updateserver-tenant-scope-spec.md).
+// Pins updateServer's CURRENT contract — it updates a row by `id` alone, with
+// no `servers.tenantId` predicate and no not-found signal — so Slice 2's
+// tenant-scoped predicate change has a proven "before" to diff against.
+type ServerRow = {
+  id: string;
+  tenantId: string;
+  name: string;
+  url: string;
+  token: string;
+  tokenIv: string;
+  authMode: string;
+  lastConnectedAt: number | null;
+  createdAt: number;
+  updatedAt: number;
+};
+
+// Fixture: two tenants, one server each, distinct ids — the same shape
+// Slice 2's cross-tenant regression test will reuse.
+function seedRows(): ServerRow[] {
+  return [
+    {
+      id: 'server-a',
+      tenantId: 'tenant-a',
+      name: 'Tenant A host',
+      url: 'https://a.example.com',
+      token: '',
+      tokenIv: '',
+      authMode: 'token',
+      lastConnectedAt: null,
+      createdAt: 1,
+      updatedAt: 1,
+    },
+    {
+      id: 'server-b',
+      tenantId: 'tenant-b',
+      name: 'Tenant B host',
+      url: 'https://b.example.com',
+      token: '',
+      tokenIv: '',
+      authMode: 'token',
+      lastConnectedAt: null,
+      createdAt: 1,
+      updatedAt: 1,
+    },
+  ];
+}
+
+let updateServerRows: ServerRow[];
+
+function makeUpdateServerDb() {
+  return {
+    update: () => ({
+      set: (patch: Partial<ServerRow>) => ({
+        where: (whereFn: (r: Record<string, unknown>) => boolean) => {
+          updateServerRows = updateServerRows.map((r) =>
+            whereFn(r as unknown as Record<string, unknown>) ? { ...r, ...patch } : r,
+          );
+          return Promise.resolve(undefined);
+        },
+      }),
+    }),
+  };
+}
+
+const updateServerCtxFor = (tenantId: string) => ({ db: makeUpdateServerDb() as never, tenantId });
+
+describe('updateServer — Slice 1 baseline (pre tenant-scope)', () => {
+  beforeEach(() => {
+    updateServerRows = seedRows();
+  });
+
+  it('same-tenant update: patches the caller tenant’s own server row', async () => {
+    await updateServer(updateServerCtxFor('tenant-a'), 'server-a', { name: 'Renamed A' });
+
+    const a = updateServerRows.find((r) => r.id === 'server-a');
+    const b = updateServerRows.find((r) => r.id === 'server-b');
+    expect(a?.name).toBe('Renamed A');
+    expect(a?.updatedAt).toBe(1_700_000_000_000);
+    expect(b).toEqual(seedRows()[1]);
+  });
+
+  it('unknown id: resolves undefined (no not-found signal) and mutates nothing', async () => {
+    await expect(
+      updateServer(updateServerCtxFor('tenant-a'), 'does-not-exist', { name: 'X' }),
+    ).resolves.toBeUndefined();
+    expect(updateServerRows).toEqual(seedRows());
   });
 });
